+++ tools/integration-test/src/tests/ordered_channel.rs
@@ -1,6 +1,9 @@
+use ibc_relayer::keyring::Store;
 use ibc_test_framework::ibc::denom::derive_ibc_denom;
 use ibc_test_framework::prelude::*;
 use ibc_test_framework::util::random::random_u64_range;
+
+const TRANSFER_COUNT: u64 = 2;
 
 #[test]
 fn test_ordered_channel() -> Result<(), Error> {
@@ -13,13 +16,19 @@ impl TestOverrides for OrderedChannelTest {
     fn modify_relayer_config(&self, config: &mut Config) {
         // Disabling clear_on_start should make the relayer not
         // relay any packet it missed before starting.
-        config.mode.packets.clear_on_start = false;
-        config.mode.packets.clear_interval = 0;
+        // config.mode.packets.clear_on_start = false;
+        // config.mode.packets.clear_interval = 0;
+
+        for mut chain in config.chains.iter_mut() {
+            // We currently need the Test key store type to persist the keys
+            // across relayer driver forks.
+            chain.key_store_type = Store::Test;
+        }
     }
 
-    fn channel_order(&self) -> Order {
-        Order::Ordered
-    }
+    // fn channel_order(&self) -> Order {
+    //     Order::Ordered
+    // }
 }
 
 impl BinaryChannelTest for OrderedChannelTest {
@@ -43,11 +52,11 @@ impl BinaryChannelTest for OrderedChannelTest {
         let amount1 = random_u64_range(1000, 5000);
 
         info!(
-            "Performing IBC transfer with amount {}, which should be relayed because its an ordered channel",
-            amount1
+            "Performing IBC transfer with amount {} times {}, which should be relayed because its an ordered channel",
+            amount1, TRANSFER_COUNT
         );
 
-        for _ in 0..5 {
+        for _ in 0..TRANSFER_COUNT {
             chains.node_a.chain_driver().transfer_token(
                 &channel.port_a.as_ref(),
                 &channel.channel_id_a.as_ref(),
@@ -58,8 +67,27 @@ impl BinaryChannelTest for OrderedChannelTest {
             )?;
         }
 
+        // Fork the relayer driver with different keys, so that we can spawn
+        // two racing relayers.
+        let relayer2 = {
+            let chain_id_a = chains.node_a.chain_id().into_value();
+            let chain_id_b = chains.node_b.chain_id().into_value();
+            let relayer_wallet_a = chains.node_a.wallets().relayer2().id().cloned_value().0;
+            let relayer_wallet_b = chains.node_b.wallets().relayer2().id().cloned_value().0;
+
+            relayer.fork(|config| {
+                for mut chain in config.chains.iter_mut() {
+                    if &chain.id == chain_id_a {
+                        chain.key_name = relayer_wallet_a.clone();
+                    } else if &chain.id == chain_id_b {
+                        chain.key_name = relayer_wallet_b.clone();
+                    }
+                }
+            })
+        };
+
         let _relayer1 = relayer.spawn_supervisor()?;
-        let _relayer2 = relayer.spawn_supervisor()?;
+        // let _relayer2 = relayer2.spawn_supervisor()?;
 
         sleep(Duration::from_secs(3));
 
@@ -90,14 +118,14 @@ impl BinaryChannelTest for OrderedChannelTest {
         // Wallet on chain A should have both amount deducted.
         chains.node_a.chain_driver().assert_eventual_wallet_amount(
             &wallet_a.address(),
-            balance_a - amount1 - amount2,
+            balance_a - (amount1 * TRANSFER_COUNT) - amount2,
             &denom_a,
         )?;
 
         // Wallet on chain B should receive both IBC transfers
         chains.node_b.chain_driver().assert_eventual_wallet_amount(
             &wallet_b.address(),
-            amount1 + amount2,
+            (amount1 * TRANSFER_COUNT) + amount2,
             &denom_b.as_ref(),
         )?;
